refactor!: move compliance overview to a dedicated recap page (#1444)
* wip

* wip

* wip

* checkpoint

* ready for review
