+++ frontend/messages/fr.json
@@ -1140,5 +1140,7 @@
 	"maxRadius": "Rayon maximal",
 	"greenZoneRadius": "Rayon de la zone verte",
 	"yellowZoneRadius": "Rayon de la zone jaune",
-	"redZoneRadius": "Rayon de la zone rouge"
+	"redZoneRadius": "Rayon de la zone rouge",
+	"recap": "Recap",
+	"sectionMoved": "Section déplacée ici"
 }
