+++ frontend/src/routes/(app)/(internal)/recap/+layout.svelte
@@ -1,0 +1,3 @@
+<div class="card bg-white shadow-lg">
+	<slot />
+</div>
